+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added WaterDrop gem with default configuration
 - Refactoring of config logic to simplify adding new dependencies that need to be configured based on #setup data
 - Gem dump
+- #37 - warn log for failed parsing of a message
 
 ## 0.3.2
 - Karafka::Params::Params lazy load merge keys with string/symbol names priorities fix